rename packages, fix structure
