+++ app/src/main/AndroidManifest.xml
@@ -51,7 +51,7 @@
             android:exported="false"
             android:theme="@style/Theme.Theme" />
         <service
-            android:name=".utils.AdbPairingService"
+            android:name="org.osservatorionessuno.cadb.AdbPairingService"
             android:exported="false"
             android:foregroundServiceType="specialUse" />
