+++ app/src/main/java/org/osservatorionessuno/cadb/AdbPairingService.java
@@ -1,4 +1,4 @@
-package org.osservatorionessuno.bugbane.utils;
+package org.osservatorionessuno.cadb;
 
 import android.annotation.SuppressLint;
 import android.app.Notification;
